Miner config, mining logic
