--- AElf.Kernel/MIner/IMiner.cs
@@ -1,8 +1,0 @@
-﻿namespace AElf.Kernel.MIner
-{
-    public interface IMiner
-    {
-        void Start();
-        void Stop();
-    }
-}
