--- AElf.Kernel/MIner/MinerConifg.cs
@@ -1,11 +1,0 @@
-﻿namespace AElf.Kernel.MIner
-{
-    public class MinerConifg : IMinerConfig
-    {
-        public Hash CoinBase { get; set; }
-        public bool IsParallel { get; set; }
-        public Hash ChainId { get; set; }
-        public ulong TxCountLimit { get; set; }
-        
-    }
-}
